# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/unidades-saude": {
      "post": {
        "description": "Cria uma nova unidade de saúde com todos os dados relacionados (horários, dados bancários, CNAEs)",
        "operationId": "UnidadeSaudeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnidadeSaudeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Unidade de saúde criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnidadeSaudeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "409": {
            "description": "CNPJ já cadastrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar nova unidade de saúde",
        "tags": [
          "Unidades de Saúde"
        ]
      },
      "get": {
        "description": "Lista todas as unidades de saúde com paginação e filtros opcionais",
        "operationId": "UnidadeSaudeController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limite de registros por página (default: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Busca por nome, fantasia, CNPJ ou razão social",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "required": false,
            "in": "query",
            "description": "Filtrar por status ativo/inativo",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "cidade",
            "required": false,
            "in": "query",
            "description": "Filtrar por cidade",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "description": "Filtrar por estado (UF)",
            "schema": {
              "example": "SP",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de unidades de saúde",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UnidadeSaudeListResponseDto"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 100
                    },
                    "page": {
                      "type": "number",
                      "example": 1
                    },
                    "limit": {
                      "type": "number",
                      "example": 10
                    },
                    "totalPages": {
                      "type": "number",
                      "example": 10
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar unidades de saúde",
        "tags": [
          "Unidades de Saúde"
        ]
      }
    },
    "/api/v1/unidades-saude/ativas": {
      "get": {
        "description": "Lista apenas as unidades de saúde ativas (para dropdowns/selects)",
        "operationId": "UnidadeSaudeController_listActive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de unidades ativas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "nomeUnidade": {
                        "type": "string"
                      },
                      "nomeFantasia": {
                        "type": "string"
                      },
                      "cnpj": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar unidades ativas",
        "tags": [
          "Unidades de Saúde"
        ]
      }
    },
    "/api/v1/unidades-saude/cidade/{cidade}": {
      "get": {
        "description": "Lista todas as unidades de saúde ativas de uma cidade específica",
        "operationId": "UnidadeSaudeController_findByCidade",
        "parameters": [
          {
            "name": "cidade",
            "required": true,
            "in": "path",
            "description": "Nome da cidade",
            "schema": {
              "example": "São Roque",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de unidades da cidade",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnidadeSaudeResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar unidades por cidade",
        "tags": [
          "Unidades de Saúde"
        ]
      }
    },
    "/api/v1/unidades-saude/cnpj/{cnpj}": {
      "get": {
        "description": "Busca uma unidade de saúde específica pelo CNPJ",
        "operationId": "UnidadeSaudeController_findByCnpj",
        "parameters": [
          {
            "name": "cnpj",
            "required": true,
            "in": "path",
            "description": "CNPJ da unidade (apenas números)",
            "schema": {
              "example": "12345678000190",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unidade de saúde encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnidadeSaudeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Unidade não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar unidade por CNPJ",
        "tags": [
          "Unidades de Saúde"
        ]
      }
    },
    "/api/v1/unidades-saude/{id}": {
      "get": {
        "description": "Busca uma unidade de saúde específica com todos os dados relacionados",
        "operationId": "UnidadeSaudeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID da unidade de saúde",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unidade de saúde encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnidadeSaudeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Unidade não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar unidade por ID",
        "tags": [
          "Unidades de Saúde"
        ]
      },
      "patch": {
        "description": "Atualiza os dados de uma unidade de saúde existente",
        "operationId": "UnidadeSaudeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID da unidade de saúde",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUnidadeSaudeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unidade de saúde atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnidadeSaudeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Unidade não encontrada"
          },
          "409": {
            "description": "CNPJ já cadastrado em outra unidade"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar unidade de saúde",
        "tags": [
          "Unidades de Saúde"
        ]
      },
      "delete": {
        "description": "Remove (soft delete) uma unidade de saúde",
        "operationId": "UnidadeSaudeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID da unidade de saúde",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Unidade removida com sucesso"
          },
          "404": {
            "description": "Unidade não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover unidade de saúde",
        "tags": [
          "Unidades de Saúde"
        ]
      }
    },
    "/api/v1/unidades-saude/{id}/ativar": {
      "patch": {
        "description": "Ativa uma unidade de saúde previamente desativada",
        "operationId": "UnidadeSaudeController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID da unidade de saúde",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unidade ativada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnidadeSaudeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Unidade não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ativar unidade de saúde",
        "tags": [
          "Unidades de Saúde"
        ]
      }
    },
    "/api/v1/unidades-saude/{id}/desativar": {
      "patch": {
        "description": "Desativa uma unidade de saúde ativa",
        "operationId": "UnidadeSaudeController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID da unidade de saúde",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unidade desativada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnidadeSaudeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Unidade não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desativar unidade de saúde",
        "tags": [
          "Unidades de Saúde"
        ]
      }
    }
  },
  "info": {
    "title": "ERP Laboratório API",
    "description": "Sistema ERP modular para laboratórios de análises clínicas e imagens",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCnaeSecundarioDto": {
        "type": "object",
        "properties": {
          "codigoCnae": {
            "type": "string",
            "example": "8621601",
            "description": "Código CNAE secundário"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição do CNAE"
          }
        },
        "required": [
          "codigoCnae"
        ]
      },
      "CreateUnidadeSaudeDto": {
        "type": "object",
        "properties": {
          "nomeUnidade": {
            "type": "string",
            "example": "Laboratório Central",
            "description": "Nome da unidade de saúde"
          },
          "codigoInterno": {
            "type": "string",
            "example": "LAB001",
            "description": "Código interno da unidade"
          },
          "cnpj": {
            "type": "string",
            "example": "12345678000190",
            "description": "CNPJ da unidade"
          },
          "razaoSocial": {
            "type": "string",
            "example": "Laboratório ABC Ltda",
            "description": "Razão social"
          },
          "nomeFantasia": {
            "type": "string",
            "example": "Lab ABC",
            "description": "Nome fantasia"
          },
          "inscricaoMunicipal": {
            "type": "string",
            "example": "123456",
            "description": "Inscrição municipal"
          },
          "inscricaoEstadual": {
            "type": "string",
            "example": "789456123",
            "description": "Inscrição estadual"
          },
          "cnes": {
            "type": "string",
            "example": "1234567",
            "description": "Código CNES"
          },
          "contatosUnidade": {
            "type": "string",
            "example": "(11) 98765-4321",
            "description": "Contatos da unidade"
          },
          "email": {
            "type": "string",
            "example": "contato@lab.com.br",
            "description": "E-mail da unidade"
          },
          "codigoServicoPrincipal": {
            "type": "string",
            "example": "001",
            "description": "Código do serviço principal"
          },
          "codigoServicoSecundario": {
            "example": [
              "002",
              "003",
              "004"
            ],
            "description": "Códigos dos serviços secundários",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cnaePrincipal": {
            "type": "string",
            "example": "8621601",
            "description": "CNAE principal"
          },
          "cep": {
            "type": "string",
            "example": "01234567",
            "description": "CEP"
          },
          "rua": {
            "type": "string",
            "example": "Rua das Flores",
            "description": "Logradouro"
          },
          "numero": {
            "type": "string",
            "example": "123",
            "description": "Número"
          },
          "bairro": {
            "type": "string",
            "example": "Centro",
            "description": "Bairro"
          },
          "complemento": {
            "type": "string",
            "example": "Sala 10",
            "description": "Complemento"
          },
          "estado": {
            "type": "string",
            "example": "SP",
            "description": "Estado (UF)"
          },
          "cidade": {
            "type": "string",
            "example": "São Paulo",
            "description": "Cidade"
          },
          "nomeResponsavel": {
            "type": "string",
            "example": "Dr. João Silva",
            "description": "Nome do responsável"
          },
          "contatoResponsavel": {
            "type": "string",
            "example": "(11) 98765-4321",
            "description": "Contato do responsável"
          },
          "emailResponsavel": {
            "type": "string",
            "example": "responsavel@lab.com.br",
            "description": "E-mail do responsável"
          },
          "irrfPercentual": {
            "type": "number",
            "example": 1.5,
            "description": "Percentual IRRF",
            "minimum": 0,
            "maximum": 100
          },
          "pisPercentual": {
            "type": "number",
            "example": 0.65,
            "description": "Percentual PIS",
            "minimum": 0,
            "maximum": 100
          },
          "cofinsPercentual": {
            "type": "number",
            "example": 3,
            "description": "Percentual COFINS",
            "minimum": 0,
            "maximum": 100
          },
          "csllPercentual": {
            "type": "number",
            "example": 1,
            "description": "Percentual CSLL",
            "minimum": 0,
            "maximum": 100
          },
          "issPercentual": {
            "type": "number",
            "example": 5,
            "description": "Percentual ISS",
            "minimum": 0,
            "maximum": 100
          },
          "ibsPercentual": {
            "type": "number",
            "example": 0,
            "description": "Percentual IBS",
            "minimum": 0,
            "maximum": 100
          },
          "cbsPercentual": {
            "type": "number",
            "example": 0,
            "description": "Percentual CBS",
            "minimum": 0,
            "maximum": 100
          },
          "reterIss": {
            "type": "boolean",
            "default": false,
            "description": "Reter ISS"
          },
          "reterIr": {
            "type": "boolean",
            "default": false,
            "description": "Reter IR"
          },
          "reterPcc": {
            "type": "boolean",
            "default": false,
            "description": "Reter PCC"
          },
          "reterIbs": {
            "type": "boolean",
            "default": false,
            "description": "Reter IBS"
          },
          "reterCbs": {
            "type": "boolean",
            "default": false,
            "description": "Reter CBS"
          },
          "optanteSimplesNacional": {
            "type": "boolean",
            "default": false,
            "description": "Optante pelo Simples Nacional"
          },
          "certificadoDigitalVinculado": {
            "type": "boolean",
            "default": false,
            "description": "Certificado digital vinculado"
          },
          "certificadoDigitalPath": {
            "type": "string",
            "description": "Caminho do arquivo do certificado digital"
          },
          "certificadoDigitalSenha": {
            "type": "string",
            "description": "Senha do certificado digital"
          },
          "certificadoDigitalValidade": {
            "format": "date-time",
            "type": "string",
            "description": "Data de validade do certificado digital"
          },
          "horariosAtendimento": {
            "description": "Horários de atendimento",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateHorarioAtendimentoDto"
            }
          },
          "dadosBancarios": {
            "description": "Dados bancários",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateDadoBancarioDto"
            }
          },
          "cnaeSecundarios": {
            "description": "CNAEs secundários",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCnaeSecundarioDto"
            }
          },
          "ativo": {
            "type": "boolean",
            "default": true,
            "description": "Status ativo/inativo"
          }
        },
        "required": [
          "nomeUnidade",
          "cnpj",
          "razaoSocial",
          "nomeFantasia"
        ]
      },
...
      "CnaeSecundarioResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "codigoCnae": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "codigoCnae",
          "ativo"
        ]
      },
      "UnidadeSaudeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nomeUnidade": {
            "type": "string"
          },
          "codigoInterno": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "razaoSocial": {
            "type": "string"
          },
          "nomeFantasia": {
            "type": "string"
          },
          "inscricaoMunicipal": {
            "type": "string"
          },
          "inscricaoEstadual": {
            "type": "string"
          },
          "cnes": {
            "type": "string"
          },
          "contatosUnidade": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "codigoServicoPrincipal": {
            "type": "string"
          },
          "codigoServicoSecundario": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cnaePrincipal": {
            "type": "string"
          },
          "cep": {
            "type": "string"
          },
          "rua": {
            "type": "string"
          },
          "numero": {
            "type": "string"
          },
          "bairro": {
            "type": "string"
          },
          "complemento": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "cidade": {
            "type": "string"
          },
          "nomeResponsavel": {
            "type": "string"
          },
          "contatoResponsavel": {
            "type": "string"
          },
          "emailResponsavel": {
            "type": "string"
          },
          "irrfPercentual": {
            "type": "number"
          },
          "pisPercentual": {
            "type": "number"
          },
          "cofinsPercentual": {
            "type": "number"
          },
          "csllPercentual": {
            "type": "number"
          },
          "issPercentual": {
            "type": "number"
          },
          "ibsPercentual": {
            "type": "number"
          },
          "cbsPercentual": {
            "type": "number"
          },
          "reterIss": {
            "type": "boolean"
          },
          "reterIr": {
            "type": "boolean"
          },
          "reterPcc": {
            "type": "boolean"
          },
          "reterIbs": {
            "type": "boolean"
          },
          "reterCbs": {
            "type": "boolean"
          },
          "optanteSimplesNacional": {
            "type": "boolean"
          },
          "certificadoDigitalVinculado": {
            "type": "boolean"
          },
          "certificadoDigitalValidade": {
            "format": "date-time",
            "type": "string"
          },
          "horariosAtendimento": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HorarioAtendimentoResponseDto"
            }
          },
          "dadosBancarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DadoBancarioResponseDto"
            }
          },
          "cnaeSecundarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CnaeSecundarioResponseDto"
            }
          },
          "ativo": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nomeUnidade",
          "cnpj",
          "razaoSocial",
          "nomeFantasia",
          "reterIss",
          "reterIr",
          "reterPcc",
          "reterIbs",
          "reterCbs",
          "optanteSimplesNacional",
          "certificadoDigitalVinculado",
          "horariosAtendimento",
          "dadosBancarios",
          "cnaeSecundarios",
          "ativo",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}